pyproject.toml
[project]
name = "cuttleagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "fastapi>=0.115.0",
    "langchain>=1.2.14",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "typer>=0.24.1",
    "uvicorn>=0.35.0",
]
